avoid anti pattern (force update) (#192)
avoid the anti pattern of artificially updating child components by finding another possibility to move key binidngs back to loopicon: just pass the position on the screen as prop
